Refactor voting and pauta management to improve separation.
Extracted Pauta-related operations into a dedicated PautaController and PautaService. Cleaned up redundant methods and logic across services to streamline responsibilities. Updated API context path for better versioning.
