+++ src/main/resources/application.properties
@@ -23,3 +23,5 @@ springdoc.api-docs.path=/api-docs
 springdoc.swagger-ui.path=/swagger-ui.html
 
 server.port=8080
+
+server.servlet.context-path=/api/v1/votacao
